refactor: SplashActivity와 LoginActivity 분리
