+++ android/app/src/main/java/com/mulberry/ody/presentation/login/LoginActivity.kt
@@ -1,13 +1,9 @@
 package com.mulberry.ody.presentation.login
 
-import android.animation.ObjectAnimator
 import android.content.Context
 import android.content.Intent
 import android.os.Bundle
 import androidx.activity.viewModels
-import androidx.core.animation.doOnEnd
-import androidx.core.splashscreen.SplashScreen
-import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
 import com.mulberry.ody.R
 import com.mulberry.ody.databinding.ActivityLoginBinding
 import com.mulberry.ody.presentation.common.binding.BindingActivity
@@ -21,26 +17,10 @@ class LoginActivity : BindingActivity<ActivityLoginBinding>(R.layout.activity_lo
         )
     }
 
-    private lateinit var splashScreen: SplashScreen
-
     override fun onCreate(savedInstanceState: Bundle?) {
-        splashScreen = installSplashScreen()
-        startSplash()
         super.onCreate(savedInstanceState)
         initializeObserve()
         viewModel.checkIfLogined()
-    }
-
-    private fun startSplash() {
-        splashScreen.setOnExitAnimationListener { splashScreenView ->
-            ObjectAnimator.ofPropertyValuesHolder(splashScreenView.iconView).run {
-                duration = SPLASH_DELAY_DURATION
-                doOnEnd {
-                    splashScreenView.remove()
-                }
-                start()
-            }
-        }
     }
 
     override fun initializeBinding() {
@@ -51,7 +31,6 @@ class LoginActivity : BindingActivity<ActivityLoginBinding>(R.layout.activity_lo
         viewModel.navigateAction.observe(this) {
             val intent = MeetingsActivity.getIntent(this@LoginActivity)
             startActivity(intent)
-            finish()
         }
         viewModel.networkErrorEvent.observe(this) {
             showRetrySnackBar { viewModel.retryLastAction() }
@@ -62,8 +41,6 @@ class LoginActivity : BindingActivity<ActivityLoginBinding>(R.layout.activity_lo
     }
 
     companion object {
-        private const val SPLASH_DELAY_DURATION = 1500L
-
         fun getIntent(context: Context): Intent {
             return Intent(context, LoginActivity::class.java)
         }
